// web/src/api/generated/model/bookmarkOut.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Revisit API
 * OpenAPI spec version: 0.1.0
 */

export interface BookmarkOut {
  createdAt: string;
  description?: string | null;
  favicon?: string | null;
  folder?: string | null;
  id: string;
  shows_preview?: boolean;
  tags?: string[];
  thumbnail?: string | null;
  title: string;
  updatedAt: string;
  url: string;
}